wikiconv/corpus-uploader/requirements.txt
requests
google-cloud-storage
absl-py
